Update TriphasicPump.py
Changed values for the remaining steps formula too.  
'self.step_count = int(round((math.sqrt(volume_value) * self.pulse_per_revolution * self.tube_diameter) / (855), 1) / 1.78) * self.gear_ratio'  
Magic number being '855' and now times the gear ratio, simple stuff.  
However, 98% stroke volume is correct and seems to be the full remaining degrees step count (Correct angle given)  
At 50% the angle is not half the step count visually.  
Around 30% achieves 50% of the remaining degrees.   
For example with no gear ratio (ratio equal to 1) for 45 degrees full run 1/2 or 22.5 degrees is achieved at 30% (not 50%).  
I have a solution; I just need the exact total degrees of the run from start sensor to squeeze plate. Then minus the remaining degrees (degrees to target) and we are left with the degrees to squeeze plate which contains the target.  
Then if there is still a problem at 50% it is in Sato's sinewave code. [Not looking forward to fixing that]  
Or create a formula to offset the stroke volume to correct value, [not ideal but shouldn't be hard]